+++ tests/Api/General/AccountTest.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Mailtrap\Tests\Api;
+namespace Mailtrap\Tests\Api\General;
 
 use Mailtrap\Api\AbstractApi;
-use Mailtrap\Api\AbstractAccount;
+use Mailtrap\Api\General\Account;
 use Mailtrap\Exception\HttpClientException;
 use Mailtrap\Helper\ResponseHelper;
 use Mailtrap\Tests\MailtrapTestCase;
 use Nyholm\Psr7\Response;
 
 /**
- * @covers AbstractAccount
+ * @covers Account
  *
  * Class AccountTest
  */
@@ -27,7 +27,7 @@ class AccountTest extends MailtrapTestCase
     {
         parent::setUp();
 
-        $this->account = $this->getMockBuilder(AbstractAccount::class)
+        $this->account = $this->getMockBuilder(Account::class)
             ->onlyMethods(['get'])
             ->setConstructorArgs([$this->getConfigMock()])
             ->getMock()