+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 2023-04-07
+
+- Breaking changes:
+    - move `accounts()` functions from the `sandbox & sending` layers to the `general`
+- Add the new `general` layer to mailtrapClient
+
 ## [1.0.0] - 2023-03-28
 
 - The initial release of the official mailtrap.io PHP client